[all components] Split portal node hook from focus portal
