+++ packages/react/src/floating-ui-react/components/FloatingPortal.tsx
@@ -1,10 +1,6 @@
 'use client';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { isNode } from '@floating-ui/utils/dom';
-import { useId } from '@base-ui/utils/useId';
-import { useIsoLayoutEffect } from '@base-ui/utils/useIsoLayoutEffect';
-import { useStableCallback } from '@base-ui/utils/useStableCallback';
 import { FocusGuard } from '../../utils/FocusGuard';
 import {
   enableFocusInside,
@@ -15,139 +11,14 @@ import {
 } from '../utils/tabbable';
 import { createChangeEventDetails } from '../../utils/createBaseUIEventDetails';
 import { REASONS } from '../../utils/reasons';
-import { createAttribute } from '../utils/createAttribute';
-import {
-  useRenderElement,
-  type UseRenderElementComponentProps,
-} from '../../utils/useRenderElement';
-import { EMPTY_OBJECT, ownerVisuallyHidden } from '../../utils/constants';
+import { ownerVisuallyHidden } from '../../utils/constants';
 import type { BaseUIComponentProps } from '../../utils/types';
-
-type FocusManagerState = null | {
-  modal: boolean;
-  open: boolean;
-  onOpenChange(
-    open: boolean,
-    data?: { reason?: string | undefined; event?: Event | undefined },
-  ): void;
-  domReference: Element | null;
-  closeOnFocusOut: boolean;
-};
-
-const PortalContext = React.createContext<null | {
-  portalNode: HTMLElement | null;
-  setFocusManagerState: React.Dispatch<React.SetStateAction<FocusManagerState>>;
-  beforeInsideRef: React.RefObject<HTMLSpanElement | null>;
-  afterInsideRef: React.RefObject<HTMLSpanElement | null>;
-  beforeOutsideRef: React.RefObject<HTMLSpanElement | null>;
-  afterOutsideRef: React.RefObject<HTMLSpanElement | null>;
-}>(null);
-
-export const usePortalContext = () => React.useContext(PortalContext);
-
-const attr = createAttribute('portal');
-
-export interface UseFloatingPortalNodeProps {
-  ref?: React.Ref<HTMLDivElement> | undefined;
-  container?:
-    | HTMLElement
-    | ShadowRoot
-    | null
-    | React.RefObject<HTMLElement | ShadowRoot | null>
-    | undefined;
-  componentProps?: UseRenderElementComponentProps<any> | undefined;
-  elementProps?: React.HTMLAttributes<HTMLDivElement> | undefined;
-}
-
-export interface UseFloatingPortalNodeResult {
-  portalNode: HTMLElement | null;
-  portalSubtree: React.ReactPortal | null;
-}
-
-export function useFloatingPortalNode(
-  props: UseFloatingPortalNodeProps = {},
-): UseFloatingPortalNodeResult {
-  const { ref, container: containerProp, componentProps = EMPTY_OBJECT, elementProps } = props;
-
-  const uniqueId = useId();
-  const portalContext = usePortalContext();
-  const parentPortalNode = portalContext?.portalNode;
-
-  const [containerElement, setContainerElement] = React.useState<HTMLElement | ShadowRoot | null>(
-    null,
-  );
-  const [portalNode, setPortalNode] = React.useState<HTMLElement | null>(null);
-  const setPortalNodeRef = useStableCallback((node: HTMLElement | null) => {
-    if (node !== null) {
-      // the useIsoLayoutEffect below watching containerProp / parentPortalNode
-      // sets setPortalNode(null) when the container becomes null or changes.
-      // So even though the ref callback now ignores null, the portal node still gets cleared.
-      setPortalNode(node);
-    }
-  });
-
-  const containerRef = React.useRef<HTMLElement | ShadowRoot | null>(null);
-
-  useIsoLayoutEffect(() => {
-    // Wait for the container to be resolved if explicitly `null`.
-    if (containerProp === null) {
-      if (containerRef.current) {
-        containerRef.current = null;
-        setPortalNode(null);
-        setContainerElement(null);
-      }
-      return;
-    }
-
-    // React 17 does not use React.useId().
-    if (uniqueId == null) {
-      return;
-    }
-
-    const resolvedContainer =
-      (containerProp && (isNode(containerProp) ? containerProp : containerProp.current)) ??
-      parentPortalNode ??
-      document.body;
-
-    if (resolvedContainer == null) {
-      if (containerRef.current) {
-        containerRef.current = null;
-        setPortalNode(null);
-        setContainerElement(null);
-      }
-      return;
-    }
-
-    if (containerRef.current !== resolvedContainer) {
-      containerRef.current = resolvedContainer;
-      setPortalNode(null);
-      setContainerElement(resolvedContainer);
-    }
-  }, [containerProp, parentPortalNode, uniqueId]);
-
-  const portalElement = useRenderElement('div', componentProps, {
-    ref: [ref, setPortalNodeRef],
-    props: [
-      {
-        id: uniqueId,
-        [attr]: '',
-      },
-      elementProps,
-    ],
-  });
-
-  // This `createPortal` call injects `portalElement` into the `container`.
-  // Another call inside `FloatingPortal`/`FloatingPortalLite` then injects the children into `portalElement`.
-  const portalSubtree =
-    containerElement && portalElement
-      ? ReactDOM.createPortal(portalElement, containerElement)
-      : null;
-
-  return {
-    portalNode,
-    portalSubtree,
-  };
-}
+import {
+  PortalContext,
+  type PortalFocusManagerState,
+  type UseFloatingPortalNodeProps,
+  useFloatingPortalNode,
+} from './useFloatingPortalNode';
 
 /**
  * Portals the floating element into a given container element — by default,
@@ -177,7 +48,7 @@ export const FloatingPortal = React.forwardRef(function FloatingPortal(
   const beforeInsideRef = React.useRef<HTMLSpanElement>(null);
   const afterInsideRef = React.useRef<HTMLSpanElement>(null);
 
-  const [focusManagerState, setFocusManagerState] = React.useState<FocusManagerState>(null);
+  const [focusManagerState, setFocusManagerState] = React.useState<PortalFocusManagerState>(null);
   const focusInsideDisabledRef = React.useRef(false);
 
   const modal = focusManagerState?.modal;
